Fix async preemption signal handling on CheriBSD freebsd64 Go.
When a goroutine is interrupted by async preemption, the hybrid Go
runtime injects a call to asyncPreempt in the current goroutine stack.
This is done by modifying the signal mcontext $sp, $lr and $pc
registers.
The CheriBSD freebsd64 compat layer introduces the _MC_CAP_VALID flag
in the mcontext.mc_flags field to control whether freebsd64
set_mcontext should update the register file, and invalidate all
capability general-purpose registers as a result. This operation is
only safe if the signal handler is modifying register contents in a
legacy aarch64 binary, not an hybrid one, because general-purpose
registers will never hold valid capabilities.
When updating the mcontext, the Go runtime now clears the
_MC_CAP_VALID flag.